+++ util.h
@@ -33,7 +33,9 @@ inline void flip_atlas(Atlas& src, Atlas& dst) {
 
 inline void line(const Camera& camera, int x1, int y1, int x2, int y2) {
     const Vector2& pos_camera = camera.get_position();
-    line((int)(x1 - pos_camera.x), (int)(y1 - pos_camera.y), (int)(x2 - pos_camera.x), (int)(y2 - pos_camera.y));
+    line((int)((float)x1 - pos_camera.x), (int)((float)y1 - pos_camera.y),
+         (int)((float)x2 - pos_camera.x), (int)((float)y2 - pos_camera.y)
+    );
 }
 
 inline void putimage_alpha(int dst_x, int dst_y, IMAGE* img) {
@@ -46,8 +48,8 @@ inline void putimage_alpha(int dst_x, int dst_y, IMAGE* img) {
 inline void putimage_alpha(const Camera& camera, int dst_x, int dst_y, IMAGE* img) {
     int w = img->getwidth(), h = img->getheight();
     const Vector2& pos_camera = camera.get_position();
-    AlphaBlend(GetImageHDC(GetWorkingImage()), (int)(dst_x - pos_camera.x), (int)(dst_y - pos_camera.y), w, h,
-               GetImageHDC(img), 0, 0, w, h, {AC_SRC_OVER, 0, 255, AC_SRC_ALPHA});
+    AlphaBlend(GetImageHDC(GetWorkingImage()), (int)((float)dst_x - pos_camera.x), (int)((float)dst_y - pos_camera.y),
+               w, h, GetImageHDC(img), 0, 0, w, h, {AC_SRC_OVER, 0, 255, AC_SRC_ALPHA});
 }
 
 inline void putimage_alpha(int dst_x, int dst_y, int width, int height, IMAGE* img, int src_x, int src_y) {
